Fix German translation for My_Computer.
First impression of the German user of TDE is that he is greeted with a
desktop icon on the very top of the desktop called "Speichermedien",
which is so long, that it wraps up this ugly way: "Speicherme-dien".
Which looks very unpolished and leads to a bad experience for first
timers. This comes from KDE3 times still. Most likely it was just
copy-pasted from other places.

The term "Speichermedien" is used otherwise in nearly all other places
(German), so changing it just for that icon looks inconsistent at first.
But in fact the original (English) name for it is also called "Computer"
and not "Speichermedien" for this icon.  So the gain of beeing
inconsistent is bigger because it looks more polished. If someone is
unhappy about that, he can open a ticket or join IRC to find another
solution.

Signed-off-by: Chris <xchrisx@uber.space>