Improve detection of libXss, OpenGL and libXext DPMS support.
The detection of libXss has been changed to ask the question
"where is libXss?" before asking "is XScreenSaverQueryInfo
available?" rather than the other way around. The test for
"XScreenSaverQueryInfo" also ensures that the include and
library flags from pkg-config are used.

The OpenGL test has been amended to make use of the include
paths and library flags returned by pkg-config to prevent the
test from failing when OpenGL is not in /usr.

The libXext test in kcontrol/energy/CMakeLists.txt is amended
to use the libXext include path and library flags. The paths
are also added to the include/link directories.

Signed-off-by: aneejit1 <aneejit1@gmail.com>
(cherry picked from commit 7fba97e18aa453c76130f587740f692cd267a216)