Taskbar: small/panel icon size selection.
By default the 16x16 value was hardcoded into the taskbar, which
is not correct.

With this commit the default value is TDEIcon::Small (can be
defined in TCC->LookNFeel->Icons) to retain the "classic" style.

This commit adds a "Small icons" checkbox which can be unchecked
to set the icon size to TDEIcon::Panel (for consistency with
other panel icons). Combined with "Icons only", this produces
a Plasma/Windows-like big icon taskbar.

This commit also ensures that in "Icons only" mode the icons
are centered in their button, for better looking result.

Finally, this commit fixes a typo in some macros
(TABSKAR -> TASKBAR).

Signed-off-by: Mavridis Philippe <mavridisf@gmail.com>