TWin: Active borders and snap tiling
This commit is a squash of the commits of TDE/tdebase#331.

In short, this backports some improvements to existing electric border
functionality from KDE, adds the snap tiling (or aerosnap) feature and
brings rudimentary support for active corners, which will be fully
implemented in a later PR.

The options dialog and the documentation has been updated to reflect
these changes.

Additionally, a new relevant option is introduced: an option for
restoring the original size of maximized/tiled windows when the user
starts dragging them. The option is set to be off by default,
preserving the traditional behaviour of KDE 3.x/TDE.

Last but not least, the term "electric" in relation to borders and
corners is replaced by "active" for clarity to the users.

Signed-off-by: Mavridis Philippe <mavridisf@gmail.com>