Deprecate TabColor setting
This is a confusing feature which saves the color of the first tab and makes it the default color in new sessions. As a result, it broke color schemes by overriding the default tab color.

This is not related to saving tab colors as part of a session.

Signed-off-by: Mavridis Philippe <mavridisf@gmail.com>
(cherry picked from commit 67ad6a14093bbdc024e6ed61dbe0dc31c3d6df67)