Fix unwanted window resizing. This resolves issue TDE/tde#57.
Some applications (like xfce4-terminal) try to manage their sizes by requesting the window manager a different size. The WM responds by resizing the window and the application tries once again to adjust its own size. This can lead to a repeated loop of request-resize which results in the application window to either shrink to the minimum allowed size or expand to the display size.

Signed-off-by: Michele Calgaro <michele.calgaro@yahoo.it>