Add support for assigning global shortcuts for suspend actions
The corresponding shortcuts have been added as part of the KDesktop shortucts, to complement similar options (logout, shutdown, reboot).

The shortcuts are only shown for the options that KSMServer reports via DCOP as valid.

In order to do be able to make a DCOP call from kdesktopbindings.cpp, the <dcopref.h> include had to be added to several files of the shortcuts TDECM.

Signed-off-by: Mavridis Philippe <mavridisf@gmail.com>