KEditFileType: prevent inadvertent editing of application/octet-stream
The default fallback mime type is not something that should be edited by the user except when explicitly requested.

A confirmation is shown now before allowing the user to edit this mime type.

Also, since KMimeType::mimeType(...) does NOT return a null value but the default mime type, the check for incorrect mime type failed, since it expected a null return value, and the default fallback mime type was used instead, making the user inadvertedly edit the default mime type. Incorrect mime types are now promptly recognized, and an error message is shown to the user in such cases.

Signed-off-by: Mavridis Philippe <mavridisf@gmail.com>