Fix kdesktop lock process hang / denial of service
Under specific circumstances, kdesktop_lock may fail to engage the screen
lock when requested.  When this occurs, the user is typically present to
observe and attempt relock, however the kdesktop_lock process may not respond
once the lock has failed to engage and the process has not been restarted.

Add a fallback handler for SIGHUP (lock engage) to the kdesktop_lock startup
application thread.  This ensures the process will always respond to a lock
request regardless of lock process state.

Signed-off-by: Timothy Pearson <kb9vqf@pearsoncomputing.net>