kcminit: fix initialization of multiple modules provided by a single library. This resolves issue #554.
The problem was that libinput was serving multiple modules and after initializing the first one, the others would be ignored because of a check on whether the library itself had already been initialized.
The solution is rather simple and checks not only the library name but also the module factory name. This ensures each library/factory module is correctly initialized.

Signed-off-by: Michele Calgaro <michele.calgaro@yahoo.it>