Add 15 color schemes taken from https://www.opendesktop.org.
- Jewels (8 schemes): taken from https://www.opendesktop.org/p/1123760, no license specified
- Seasons (4 schemes): taken from https://www.opendesktop.org/p/1123759, no license specified
- Different: taken from https://www.opendesktop.org/p/1123251, no license specified
- Lila: taken from https://www.opendesktop.org/p/1123580, no license specified
- Pinkie: taken from https://www.opendesktop.org/p/1994203, Creative Common 0 license

Given the files are available on opendesktop.org, it seems reasonable to assumed that all files
without explicit license are publicly available and therefore redistributable under GPL2 license.
Should this not be the case, we kindly ask the authors of the files to reach out
to us and we will promptly remove any offending file.

Signed-off-by: Michele Calgaro <michele.calgaro@yahoo.it>