kdesktop: make sure the saver/lock remains responsive even when invokation has failed.
This could happen for example when a popup menu is open and the saver
autolock timeout expires. Before this commit, both kdesktop and the
X11 server would be in a dirty start and the saver/lock would either not
work anymore or not hide the windows on the screen.
It was necessary to kill and restart kdesktop to restore proper
functionality.

Note: the conditional return value of 0 or 12 at main.cpp:554 is a temporary
measure to prevent the screen to forcibly lock if the user forgets a popup
menu open and the saver timeout expires. Once the issue with input grabbing
is solved, we can remove the conditional return value and simply return an error
in such condition. Same applies to the variable trinity_desktop_lock_failed_grab,
which will no longer be needed in such case.o

Signed-off-by: Michele Calgaro <michele.calgaro@yahoo.it>