kicker: add an option to set external margin for tray icons
This commit adds an option to set both: margin (space from an icon to
the border) and spacing (space between two neighbouring icons). The
option to tweak spacing was previously introduced in fa284a4598 and was
called "margin" the actual margin was reduced to 0. Unfortunately that
had its own unwanted side-effects listed in #647. Before fa284a4598 the
margin and spacing were fixed to be 1.

This commit restores old defaults for margin & spacing i.e. 1 to better
conform to config of older systems. Effectively this resolved most
issues raised in #647 and eliminates any regressions, though there are
some caveat cases when the behaviour is not ideal e.g. next settings:

    panel_width=47
    spacing=1
    margin=0

result in next layout:

    [0px][icon][2px][icon][1px]

when ideal would be:

    [1px][icon][1px][icon][1px]

On the other hand in such a case a user should probably just use
margin=spacing=1.

Besides that commit contains some minor refactoring.

Bug: https://mirror.git.trinitydesktop.org/gitea/TDE/tdebase/issues/647
Signed-off-by: Alexander Golubev <fatzer2@gmail.com>