Web Shortcuts: add support for groups
The web shortcuts can now be displayed in groups based on their
category (Web, Images, Development, etc.). The user can modify
the category of an item by picking one of the built-in categories
or creating custom groups by providing a different category name.

When the category groups are hidden, the user can still see
the categories appear in a third column in the list view.

This commit also updates the search engines enabled by default to
some sensible defaults and makes the minimum size of the dialog
smaller.

Signed-off-by: Philippe Mavridis <philippe.mavridis@yandex.com>