Revert Konqueror graphics to its original form
There were changes made to original Konqueror graphics in #632.  These were
questionable in that: a) they were an instance of fixing something that wasn't
broken; and b) editing was highly unskillful and came from a contributor that
also used AI to generate new wallpaper in TDE.

Some of these editing issues were later corrected in #649, but the texture still
exhibits transparency issues.

This commit reverts the texture to its original state, which had no issues that
needed fixing.

See tdeartwork#58